requirements.txt
requests==2.31.0
aiohttp
beautifulsoup4==4.12.2
urllib3==2.0.7
flask==2.3.3
flask-compress
waitress
playwright
nest-asyncio==1.5.8
psutil
bcrypt==4.1.2
markdown
python-dotenv
PyYAML
anytree
psycopg2-binary
